Filters on Paging folders update
